requirements.txt
playwright>=1.40.0
pytesseract>=0.3.10
Pillow>=10.0.0
python-telegram-bot>=20.0
python-dotenv>=1.0.0
opencv-python-headless>=4.8.0
numpy>=1.24.0
